requirements.txt
<<<<<<< HEAD
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-py==999.0.5
mysql-connector-python==8.4.0
Werkzeug==3.1.3
=======
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mysql-connector-py==999.0.5
mysql-connector-python==8.4.0
Werkzeug==3.1.3
>>>>>>> 07cd6b4f690f4d00f9a608ddd9e937c438b8e361
